+++ MatrixSDK/JSONModels/MXEvent.h
@@ -90,6 +90,7 @@ typedef NS_ENUM(NSInteger, MXEventType)
     MXEventTypeSecretSend,
     MXEventTypeSecretStorageDefaultKey,
     MXEventTypeTaggedEvents,
+    MXEventTypeSpaceChild,
 
     // The event is a custom event. Refer to its `MXEventTypeString` version
     MXEventTypeCustom = 1000
@@ -145,6 +146,7 @@ FOUNDATION_EXPORT NSString *const kMXEventTypeStringCallAssertedIdentityUnstable
 FOUNDATION_EXPORT NSString *const kMXEventTypeStringSticker;
 FOUNDATION_EXPORT NSString *const kMXEventTypeStringRoomTombStone;
 FOUNDATION_EXPORT NSString *const kMXEventTypeStringTaggedEvents;
+FOUNDATION_EXPORT NSString *const kMXEventTypeStringSpaceChild;
 
 // Interactive key verification
 FOUNDATION_EXPORT NSString *const kMXEventTypeStringKeyVerificationRequest;